requirements.txt
streamlit
openreview-py
pandas
requests
beautifulsoup4
openai
lxml
watchdog
